# implement concat without intermediate buffer