@arvidn 2021-07-04
I do think this is worth it. The RSS reduction is quite significant and one of the things I worry about a bit.